src/requirements.txt
azure-functions
google-api-python-client==1.7.11
google-auth==1.10.1
google-auth-httplib2==0.0.3
httplib2==0.16.0
oauth2client==4.1.3